requirements.txt
Flask==2.3.0
flask-cors==4.0.0
python-dotenv==1.0.0
colorama==0.4.6
gunicorn==20.1.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.130.0